Remove SMTP quota
