+++ src/Repositories/Messages/MessageTenantRepositoryInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Sendportal\Base\Repositories\Messages;
 
-use Carbon\Carbon;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
 use Sendportal\Base\Interfaces\BaseTenantInterface;
@@ -14,8 +13,6 @@ interface MessageTenantRepositoryInterface extends BaseTenantInterface
     public function paginateWithSource(int $workspaceId, string $orderBy = 'name', array $relations = [], int $paginate = 25, array $parameters = []): LengthAwarePaginator;
 
     public function recipients(int $workspaceId, string $sourceType, int $sourceId): LengthAwarePaginator;
-
-    public function countForSourcesBetween(int $workspaceId, array $sourceIds, Carbon $start, Carbon $end): int;
 
     public function opens(int $workspaceId, string $sourceType, int $sourceId): LengthAwarePaginator;
