+++ tests/Unit/Services/QuotaServiceTest.php
@@ -6,6 +6,7 @@ namespace Tests\Unit\Services;
 
 use Aws\Sdk;
 use Aws\Ses\SesClient;
+use Sendportal\Base\Exceptions\MessageLimitReachedException;
 use Sendportal\Base\Interfaces\QuotaServiceInterface;
 use Sendportal\Base\Models\Campaign;
 use Sendportal\Base\Models\EmailService;
@@ -74,172 +75,32 @@ class QuotaServiceTest extends TestCase
     }
 
     /** @test */
-    public function smtp_hourly_message_limit_reached()
+    public function a_message_limit_reached_exception_is_thrown()
     {
         // given
-        $emailService = EmailService::factory()->create(
-            [
-                'type_id' => EmailServiceType::SMTP,
-                'settings' => [
-                    'quota_limit' => 5,
-                    'quota_period' => EmailService::QUOTA_PERIOD_HOUR,
-                ],
-            ]
-        );
+        $emailService = EmailService::factory()->create(['type_id' => EmailServiceType::SES]);
 
         $campaign = Campaign::factory()->create(
             [
                 'email_service_id' => $emailService->id,
                 'workspace_id' => $emailService->workspace_id,
+                'content' => 'test',
             ]
         );
 
-        Message::factory()->count(5)->create(
+        $message = Message::factory()->create(
             [
                 'source_id' => $campaign->id,
                 'workspace_id' => $emailService->workspace_id,
-                'sent_at' => now()->subMinute(),
             ]
         );
 
-        // then
-        self::assertTrue($this->quotaService->hasReachedMessageLimit($emailService));
-    }
+        $this->mockMailAdapter(1, 1);
 
-    /** @test */
-    public function smtp_hourly_message_limit_not_reached()
-    {
-        // given
-        $emailService = EmailService::factory()->create(
-            [
-                'type_id' => EmailServiceType::SMTP,
-                'settings' => [
-                    'quota_limit' => 5,
-                    'quota_period' => EmailService::QUOTA_PERIOD_HOUR,
-                ],
-            ]
-        );
+        $this->withoutExceptionHandling()
+            ->expectException(MessageLimitReachedException::class);
 
-        $campaign = Campaign::factory()->create(
-            [
-                'email_service_id' => $emailService->id,
-                'workspace_id' => $emailService->workspace_id,
-            ]
-        );
-
-        Message::factory()->count(3)->create(
-            [
-                'source_id' => $campaign->id,
-                'workspace_id' => $emailService->workspace_id,
-                'sent_at' => now()->subHours(2),
-            ]
-        );
-
-        Message::factory()->count(2)->create(
-            [
-                'source_id' => $campaign->id,
-                'workspace_id' => $emailService->workspace_id,
-                'sent_at' => now()->subMinute(),
-            ]
-        );
-
-        // then
-        self::assertFalse($this->quotaService->hasReachedMessageLimit($emailService));
-    }
-
-    /** @test */
-    public function smtp_daily_message_limit_reached()
-    {
-        // given
-        $emailService = EmailService::factory()->create(
-            [
-                'type_id' => EmailServiceType::SMTP,
-                'settings' => [
-                    'quota_limit' => 5,
-                    'quota_period' => EmailService::QUOTA_PERIOD_DAY,
-                ],
-            ]
-        );
-
-        $campaign = Campaign::factory()->create(
-            [
-                'email_service_id' => $emailService->id,
-                'workspace_id' => $emailService->workspace_id,
-            ]
-        );
-
-        Message::factory()->count(5)->create(
-            [
-                'source_id' => $campaign->id,
-                'workspace_id' => $emailService->workspace_id,
-                'sent_at' => now()->subHours(12),
-            ]
-        );
-
-        // then
-        self::assertTrue($this->quotaService->hasReachedMessageLimit($emailService));
-    }
-
-    /** @test */
-    public function smtp_daily_message_limit_not_reached()
-    {
-        // given
-        $emailService = EmailService::factory()->create(
-            [
-                'type_id' => EmailServiceType::SMTP,
-                'settings' => [
-                    'quota_limit' => 5,
-                    'quota_period' => EmailService::QUOTA_PERIOD_DAY,
-                ],
-            ]
-        );
-
-        $campaign = Campaign::factory()->create(
-            [
-                'email_service_id' => $emailService->id,
-                'workspace_id' => $emailService->workspace_id,
-            ]
-        );
-
-        Message::factory()->count(3)->create(
-            [
-                'source_id' => $campaign->id,
-                'workspace_id' => $emailService->workspace_id,
-                'sent_at' => now()->subHours(2),
-            ]
-        );
-
-        // then
-        self::assertFalse($this->quotaService->hasReachedMessageLimit($emailService));
-    }
-
-    /** @test */
-    public function smtp_message_limit_not_set()
-    {
-        // given
-        $emailService = EmailService::factory()->create(
-            [
-                'type_id' => EmailServiceType::SMTP,
-            ]
-        );
-
-        $campaign = Campaign::factory()->create(
-            [
-                'email_service_id' => $emailService->id,
-                'workspace_id' => $emailService->workspace_id,
-            ]
-        );
-
-        Message::factory()->count(3)->create(
-            [
-                'source_id' => $campaign->id,
-                'workspace_id' => $emailService->workspace_id,
-                'sent_at' => now()->subHours(2),
-            ]
-        );
-
-        // then
-        self::assertFalse($this->quotaService->hasReachedMessageLimit($emailService));
+        $this->post(route('sendportal.messages.send'), ['id' => $message->id]);
     }
 
     protected function mockMailAdapter(int $quota = null, int $sent = 0): void
